fixed compiler warning about comparing signed to unsigned integer
compiler warning:
- [-Wsign-compare]

arcadia/unit.cpp/Unit::Study
Condition compares signed to unsigned int:
Unit::GetMen(void) returns a signed int with lowest value=0 and highest value=INT_MAX.
Skill::days and Skill::experience are unsigned int.

Fixed by casting GetMen() to unsigned int.

arcadia/unit.cpp:1252:19: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]

+++ arcadia/unit.cpp
@@ -1249,7 +1249,7 @@ int Unit::Study(int sk, int days, int quiet, int overflow)
 			s = (Skill *) skills.First();
 			if (s->type != sk) {
 			    //Real experience patch
-			    if(s->days >= 30*GetMen() || s->experience >= 30*GetMen()) {
+			    if(s->days >= 30*static_cast<unsigned int>(GetMen()) || s->experience >= 30*static_cast<unsigned int>(GetMen())) {
     				Error("STUDY: Can know only 1 skill.", quiet);
     				return 0;
 				}
